Split RuntimeData into separate file
