+++ src/runtime/Python.Runtime.csproj
@@ -137,6 +137,7 @@
     <Compile Include="pythonexception.cs" />
     <Compile Include="pytuple.cs" />
     <Compile Include="runtime.cs" />
+    <Compile Include="runtime_data.cs" />
     <Compile Include="runtime_state.cs" />
     <Compile Include="typemanager.cs" />
     <Compile Include="typemethod.cs" />
